+++ extensions/raycast-ollama/CHANGELOG.md
@@ -14,3 +14,9 @@
 - Command 'Ask Ollama' no longer use fallback for text, replaced with query.
 - Inference statistics disabled by default and can be enable from extension preferences.
 - Text is displayed realtime.
+
+## [Improvement] - 2023-07-31
+
+- New Command 'Ask Ollama with Custom Prompt' permit to use ask Ollama using a custom prompt. Use it with Quicklinks for create a custom Command.
+- New Command 'Custom Command' permit to create custom command for perform task on selected text. Use it with Quicklinks for create a custom Command.
+- Code improvement.